fix: moving getPrivateKey
